refactor : 번역 상수 폴더 이동
