+++ src/constants/translate/information/en.ts
