+++ src/constants/translate/information/index.ts
