+++ src/constants/translate/information/ko.ts
